Pushy pushy
Greetings fellow developers,

I am pleased to announce the latest commit to our project: the potato-duck hybrid feature. This revolutionary new addition combines the starchy goodness of potatoes with the savory succulence of duck confit, resulting in code that is not only twice as delicious, but also half as quackable.

However, please be advised that the implementation of this feature may cause some temporal discombobulation due to its potato-powered time travel capabilities. As such, we recommend that you buckle up and hold onto your forks as you journey through the pages of code.

Thank you for your attention and happy coding!